Implement regularized part of hypersingular Burton-Miller integral
The accuracy of the result from the test case with the fully reflective plane
is rather low and can't be improved unless the singularity in the hypersingular
part is lifted.
For this reason, a regularization part is added to the hypersingular part of the
Burton-Miller integral, which subtracts the singularity at zero.
For the result to be correct, this same regularization part has to be integrated
analytically and added to the final result.
In this commit, the accuracy of the regularized integral is checked by comparing
the adaptive integration method included in scipy with a fixed_quad integration.