Increase accuracy requirement for the fully reflective plane
Increasing the length of the reflective plane now actually improves the
solution.
With the new regularized integral formulation for Burton-Miller the relative
error can now easily be pushed below the level where it was with scipy quad.
For 400 elements and a length from -20 to +20 meters, the error can be pushed
belwo 1e3, while it stays above that level for the Kirchoff-Helmholtz
formulation were scipy quad is still used.