Remove unnecessary parameter from solver
The Kirchhoff-Helmholtz method, other than the Burton-Miller method, does not
need the gradient of the incident field at the mesh points.
To make testing easier and to have only one single interface for the solver
functions, its parameterlist was previously modified to include this unnecessary
parameter.
Since solvers should be able to have different input parameters, this
unnecessary parameter was now removed from the function.
For testing a wrapper was created that just discards the gradient of the
incident field to make their parameterlists identical.